LayoutTests/platform/chromium-mac/fast/xpath/nsresolver-exception-expected.txt

CONSOLE MESSAGE: line 14: Uncaught Error: this is an exception
SUCCESS
